refactor: add hibernate properties customizer class
