+++ mta.yaml
@@ -1,6 +1,6 @@
 _schema-version: "3.3"
 ID: mta_lms
-version: 1.4.0
+version: 1.5.0
 
 modules:
   - name: lms
