+++ src/main/java/org/example/learningsystem/core/multitenancy/config/CustomHibernatePropertiesCustomizer.java
@@ -1,0 +1,26 @@
+package org.example.learningsystem.core.multitenancy.config;
+
+import lombok.RequiredArgsConstructor;
+import org.example.learningsystem.core.multitenancy.resolver.TenantIdentifierResolver;
+import org.hibernate.engine.jdbc.connections.spi.MultiTenantConnectionProvider;
+import org.springframework.boot.autoconfigure.orm.jpa.HibernatePropertiesCustomizer;
+import org.springframework.stereotype.Component;
+
+import java.util.Map;
+
+import static org.hibernate.cfg.MultiTenancySettings.MULTI_TENANT_CONNECTION_PROVIDER;
+import static org.hibernate.cfg.MultiTenancySettings.MULTI_TENANT_IDENTIFIER_RESOLVER;
+
+@Component
+@RequiredArgsConstructor
+public class CustomHibernatePropertiesCustomizer implements HibernatePropertiesCustomizer {
+
+    private final MultiTenantConnectionProvider<String> multiTenantConnectionProvider;
+    private final TenantIdentifierResolver tenantIdentifierResolver;
+
+    @Override
+    public void customize(Map<String, Object> hibernateProperties) {
+        hibernateProperties.put(MULTI_TENANT_IDENTIFIER_RESOLVER, tenantIdentifierResolver);
+        hibernateProperties.put(MULTI_TENANT_CONNECTION_PROVIDER, multiTenantConnectionProvider);
+    }
+}
