+++ src/main/java/org/example/learningsystem/core/multitenancy/provider/LocalMultiTenantConnectionProvider.java
@@ -1,4 +1,4 @@
-package org.example.learningsystem.core.multitenancy.config;
+package org.example.learningsystem.core.multitenancy.provider;
 
 import lombok.extern.slf4j.Slf4j;
 import org.checkerframework.checker.initialization.qual.Initialized;
@@ -7,21 +7,17 @@ import org.checkerframework.checker.nullness.qual.UnknownKeyFor;
 import org.example.learningsystem.core.multitenancy.service.TenantSchemaProvider;
 import org.hibernate.engine.jdbc.connections.spi.MultiTenantConnectionProvider;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.autoconfigure.orm.jpa.HibernatePropertiesCustomizer;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
 import javax.sql.DataSource;
 import java.sql.Connection;
 import java.sql.SQLException;
-import java.util.Map;
-
-import static org.hibernate.cfg.MultiTenancySettings.MULTI_TENANT_CONNECTION_PROVIDER;
 
 @Component
 @Profile("!cloud")
 @Slf4j
-public class LocalMultiTenantConnectionProvider implements MultiTenantConnectionProvider<String>, HibernatePropertiesCustomizer {
+public class LocalMultiTenantConnectionProvider implements MultiTenantConnectionProvider<String> {
 
     private final DataSource dataSource;
     private final String defaultTenant;
@@ -75,10 +71,5 @@ public class LocalMultiTenantConnectionProvider implements MultiTenantConnection
     @Override
     public <T> T unwrap(@UnknownKeyFor @NonNull @Initialized Class<T> aClass) {
         return null;
-    }
-
-    @Override
-    public void customize(Map<String, Object> hibernateProperties) {
-        hibernateProperties.put(MULTI_TENANT_CONNECTION_PROVIDER, this);
     }
 }
