+++ src/main/java/org/example/learningsystem/core/multitenancy/resolver/TenantIdentifierResolver.java
@@ -1,18 +1,14 @@
-package org.example.learningsystem.core.multitenancy.config;
+package org.example.learningsystem.core.multitenancy.resolver;
 
 import org.example.learningsystem.core.multitenancy.util.TenantContext;
 import org.hibernate.context.spi.CurrentTenantIdentifierResolver;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.autoconfigure.orm.jpa.HibernatePropertiesCustomizer;
 import org.springframework.stereotype.Component;
 
-import java.util.Map;
-
 import static java.util.Objects.requireNonNullElse;
-import static org.hibernate.cfg.MultiTenancySettings.MULTI_TENANT_IDENTIFIER_RESOLVER;
 
 @Component
-public class TenantIdentifierResolver implements CurrentTenantIdentifierResolver<String>, HibernatePropertiesCustomizer {
+public class TenantIdentifierResolver implements CurrentTenantIdentifierResolver<String> {
 
     private final String defaultTenant;
 
@@ -29,10 +25,5 @@ public class TenantIdentifierResolver implements CurrentTenantIdentifierResolver
     @Override
     public boolean validateExistingCurrentSessions() {
         return true;
-    }
-
-    @Override
-    public void customize(Map<String, Object> hibernateProperties) {
-        hibernateProperties.put(MULTI_TENANT_IDENTIFIER_RESOLVER, this);
     }
 }
