+++ src/main/resources/application.yaml
@@ -20,6 +20,9 @@ spring:
     basename: i18n/messages
     encoding: UTF-8
 
+  profiles:
+    active: local
+
 basic-authentication:
   manager:
     username: manager@gmail.com
